Cleanup and codestyle fixes
