+++ src/GUI/Overlay/ScriptedOverlay.cpp
